Type Alias parachain_template_runtime::xcm_config::LocationToAccountId

pub type LocationToAccountId = (ParentIsPreset<AccountId>, SiblingParachainConvertsVia<Sibling, AccountId>, AccountId32Aliases<RelayNetwork, AccountId>);

Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.