Crate parachain_template_runtime

Re-exports

pub use pallet_template;

Modules

api

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

Ancestry

ExistentialDeposit

GenesisConfig

MaxDownwardMessageWeight

MaxInstructions

MaxLocks

MaxReserves

MinimumPeriod

OnRuntimeUpgrade

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParentOrParentsExecutivePlurality

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

RelayChainOrigin

RelayNetwork

ReservedDmpWeight

ReservedXcmpWeight

RocLocation

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

TransactionByteFee

Relay Chain TransactionByteFee / 10

UnitWeightCost

Version

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

XcmConfig

Enums

Call

Event

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

Constants

DAYS

EXISTENTIAL_DEPOSIT

The existential deposit. Set to 1/10 of the Rococo Relay Chain.

HOURS

MICROUNIT

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLIUNIT

MINUTES

PRIMARY_PROBABILITY

SLOT_DURATION

UNIT

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AuthorFilter

AuthorFilterConfig

AuthorInherent

Balance

Balance of an account.

Balances

BalancesConfig

Barrier

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

CumulusXcm

DmpQueue

EligibilityValue

The type of eligibility to use

Executive

Executive: handles dispatch to the various modules.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Index

Index of a transaction in the chain.

LocalAssetTransactor

Means for transacting assets on this chain.

LocalOriginToLocation

No local origins on this chain are allowed to dispatch XCM sends/executions.

LocationToAccountId

Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

NimbusId

A nimbus author identifier (A public key).

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PotentialAuthorSet

PotentialAuthorSetConfig

RandomnessCollectiveFlip

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

TemplatePallet

Timestamp

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.

XcmpQueue