Crate parachain_template_runtime

Modules

apis

configs

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

Runtime

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SessionKeys

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

DAYS

EXISTENTIAL_DEPOSIT

The existential deposit. Set to 1/10 of the Connected Relay Chain.

HOURS

MICRO_UNIT

MILLI_SECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLI_UNIT

MINUTES

PARACHAIN_ID

Parachain id used for genesis config presets of parachain template.

SLOT_DURATION

UNIT

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CollatorSelection

CollatorSelectionConfig

CumulusXcm

Executive

Executive: handles dispatch to the various modules.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

MessageQueue

Nonce

Index of a transaction in the chain.

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Session

SessionConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

Sudo

SudoConfig

System

SystemConfig

TemplatePallet

Timestamp

TransactionPayment

TransactionPaymentConfig

TxExtension

The extension to the basic transaction logic.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

WeightReclaim

XcmpQueue